Extract ContentEmitterAwareDetector to share
